+++ mockserver-core/src/test/java/org/mockserver/log/MockServerEventLogRequestLogEntryVerificationSequenceTest.java
@@ -3,6 +3,7 @@ package org.mockserver.log;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockserver.configuration.ConfigurationProperties;
 import org.mockserver.log.model.LogEntry;
 import org.mockserver.logging.MockServerLogger;
 import org.mockserver.scheduler.Scheduler;
@@ -481,6 +482,107 @@ public class MockServerEventLogRequestLogEntryVerificationSequenceTest {
                 "}, {" + NEW_LINE +
                 "  \"path\" : \"four\"" + NEW_LINE +
                 "} ]>"));
+    }
+
+    @Test
+    public void shouldFailVerificationSequenceWithLimitedReturnedRequests() {
+        Integer originalMaximumNumberOfRequestToReturnInVerificationFailure = ConfigurationProperties.maximumNumberOfRequestToReturnInVerificationFailure();
+        try {
+            // given
+            ConfigurationProperties.maximumNumberOfRequestToReturnInVerificationFailure(1);
+
+            // when
+            mockServerEventLog.add(
+                new LogEntry()
+                    .setHttpRequest(request("one"))
+                    .setType(RECEIVED_REQUEST)
+            );
+            mockServerEventLog.add(
+                new LogEntry()
+                    .setHttpRequest(request("multi"))
+                    .setType(RECEIVED_REQUEST)
+            );
+            mockServerEventLog.add(
+                new LogEntry()
+                    .setHttpRequest(request("three"))
+                    .setType(RECEIVED_REQUEST)
+            );
+            mockServerEventLog.add(
+                new LogEntry()
+                    .setHttpRequest(request("multi"))
+                    .setType(RECEIVED_REQUEST)
+            );
+            mockServerEventLog.add(
+                new LogEntry()
+                    .setHttpRequest(request("four"))
+                    .setType(RECEIVED_REQUEST)
+            );
+
+            // then - next to each other
+            assertThat(verify(
+                    new VerificationSequence()
+                        .withRequests(
+                            request("multi"),
+                            request("one")
+                        )
+                ),
+                is("Request sequence not found, expected:<[ {" + NEW_LINE +
+                    "  \"path\" : \"multi\"" + NEW_LINE +
+                    "}, {" + NEW_LINE +
+                    "  \"path\" : \"one\"" + NEW_LINE +
+                    "} ]> but was not found, found 5 other requests"));
+            assertThat(verify(
+                    new VerificationSequence()
+                        .withRequests(
+                            request("four"),
+                            request("multi")
+                        )
+                ),
+                is("Request sequence not found, expected:<[ {" + NEW_LINE +
+                    "  \"path\" : \"four\"" + NEW_LINE +
+                    "}, {" + NEW_LINE +
+                    "  \"path\" : \"multi\"" + NEW_LINE +
+                    "} ]> but was not found, found 5 other requests"));
+            // then - not next to each other
+            assertThat(verify(
+                    new VerificationSequence()
+                        .withRequests(
+                            request("three"),
+                            request("one")
+                        )
+                ),
+                is("Request sequence not found, expected:<[ {" + NEW_LINE +
+                    "  \"path\" : \"three\"" + NEW_LINE +
+                    "}, {" + NEW_LINE +
+                    "  \"path\" : \"one\"" + NEW_LINE +
+                    "} ]> but was not found, found 5 other requests"));
+            assertThat(verify(
+                    new VerificationSequence()
+                        .withRequests(
+                            request("four"),
+                            request("one")
+                        )
+                ),
+                is("Request sequence not found, expected:<[ {" + NEW_LINE +
+                    "  \"path\" : \"four\"" + NEW_LINE +
+                    "}, {" + NEW_LINE +
+                    "  \"path\" : \"one\"" + NEW_LINE +
+                    "} ]> but was not found, found 5 other requests"));
+            assertThat(verify(
+                    new VerificationSequence()
+                        .withRequests(
+                            request("four"),
+                            request("three")
+                        )
+                ),
+                is("Request sequence not found, expected:<[ {" + NEW_LINE +
+                    "  \"path\" : \"four\"" + NEW_LINE +
+                    "}, {" + NEW_LINE +
+                    "  \"path\" : \"three\"" + NEW_LINE +
+                    "} ]> but was not found, found 5 other requests"));
+        } finally {
+            ConfigurationProperties.maximumNumberOfRequestToReturnInVerificationFailure(originalMaximumNumberOfRequestToReturnInVerificationFailure);
+        }
     }
 
     @Test